Created commands/download/transform_html_to_csv_w_pandas_for_bbfiresults.py;

+++ commands/download/convert_comma_to_point_as_decimal.py
@@ -14,14 +14,16 @@ folderpath = (
 )
 
 
-def save_file(basefolderpath, previous_filename, output_text):
-  name, ext = os.path.splitext(previous_filename)
-  newname = name + '-conv' + ext
-  print('Saving file', newname)
-  filepath = os.path.join(basefolderpath, newname)
+def save_file_with_text(filepath, text):
+  filename = os.path.split(filepath)[-1]
+  if os.path.isfile(filepath):
+    print('Cannot disk-write', filename, 'for it already exists')
+    return False
+  print('Saving file', filename)
   fd = open(filepath, 'w', encoding='utf-8')
-  fd.write(output_text)
+  fd.write(text)
   fd.close()
+  return True
 
 
 def step1_convert_thousands_points_to_empty_in_text_via_regexp(input_text):
@@ -40,6 +42,49 @@ def step2_convert_comma_to_point_in_text_via_regexp(input_text):
   """
   output_text = comma_decimal_place_compiled_re.sub('.', input_text)
   return output_text
+
+
+class SingleFileConverter:
+  def __init__(self, input_filepath, output_filepath):
+    self.input_filepath = input_filepath
+    self.output_filepath = output_filepath
+    self.process()
+
+  @property
+  def input_filename(self):
+    return os.path.split(self.input_filepath)[-1]
+
+  @property
+  def output_filename(self):
+    return os.path.split(self.output_filepath)[-1]
+
+  def convert_numbers_comma_to_point_n_savefile(self):
+    print('SingleFileConverter Converting', self.input_filename, self.output_filename)
+    input_text = open(self.input_filepath).read()
+    # Step 1
+    input_text = step1_convert_thousands_points_to_empty_in_text_via_regexp(input_text)
+    # Step 2
+    output_text = step2_convert_comma_to_point_in_text_via_regexp(input_text)
+    # saves to output_filepath
+    _ = save_file_with_text(self.output_filepath, output_text)
+
+  def raise_exception_if_inputfile_is_missing(self):
+    if not os.path.isfile(self.input_filepath):
+      error_msg = 'Error: input file is missing =>' + self.input_filename
+      raise OSError(error_msg)
+
+  def outputfile_already_exists(self):
+    if os.path.isfile(self.output_filepath):
+      scrmsg = 'Output file [%s] already exists. Continuing.' % self.output_filename
+      print(scrmsg)
+      return True
+    return False
+
+  def process(self):
+    self.raise_exception_if_inputfile_is_missing()
+    if self.outputfile_already_exists():
+      return
+    self.convert_numbers_comma_to_point_n_savefile()
 
 
 class Converter:
@@ -65,7 +110,10 @@ class Converter:
     input_text = step1_convert_thousands_points_to_empty_in_text_via_regexp(input_text)
     # Step 2
     output_text = step2_convert_comma_to_point_in_text_via_regexp(input_text)
-    save_file(basefolderpath, inputfilepath, output_text)
+    name, dotext = os.path.splitext(filename)
+    output_filename = name + ' point-sep' + dotext
+    output_filepath = os.path.join(basefolderpath, output_filename)
+    _ = save_file_with_text(output_filepath, output_text)
     return output_text
 
   def process(self):
